+++ megalinter/config.py
@@ -111,6 +111,12 @@ def get(config_var=None, default=None):
     return val
 
 
+def set(config_var, value):
+    global CONFIG_DATA
+    assert CONFIG_DATA is not None, "Config has not been initialized yet !"
+    CONFIG_DATA[config_var] = value
+
+
 def get_list(config_var, default=None):
     var = get(config_var, None)
     if var is not None: