+++ megalinter/tests/test_megalinter/helpers/utilstest.py
@@ -14,11 +14,14 @@ from distutils.dir_util import copy_tree
 
 from git import Repo
 from megalinter import Megalinter, config, utils
-from megalinter.constants import DEFAULT_REPORT_FOLDER_NAME
+from megalinter.constants import (
+    DEFAULT_REPORT_FOLDER_NAME,
+    DEFAULT_DOCKER_WORKSPACE_DIR,
+)
 
 REPO_HOME = (
-    "/tmp/lint"
-    if os.path.isdir("/tmp/lint")
+    DEFAULT_DOCKER_WORKSPACE_DIR
+    if os.path.isdir(DEFAULT_DOCKER_WORKSPACE_DIR)
     else os.path.dirname(os.path.abspath(__file__))
     + os.path.sep
     + ".."
@@ -46,11 +49,11 @@ def linter_test_setup(params=None):
         "IGNORE_GITIGNORED_FILES",
         "IGNORE_GENERATED_FILES",
         "SHOW_ELAPSED_TIME",
-        "UPDATED_SOURCES_REPORTER"
+        "UPDATED_SOURCES_REPORTER",
         "MEGALINTER_FLAVOR",
         "FLAVOR_SUGGESTIONS",
         "SARIF_REPORTER",
-        "LOG_FILE"
+        "LOG_FILE",
     ]:
         if key in os.environ:
             del os.environ[key]
@@ -65,8 +68,8 @@ def linter_test_setup(params=None):
     )
     # Root path of default rules
     root_dir = (
-        "/tmp/lint"
-        if os.path.isdir("/tmp/lint")
+        DEFAULT_DOCKER_WORKSPACE_DIR
+        if os.path.isdir(DEFAULT_DOCKER_WORKSPACE_DIR)
         else os.path.relpath(
             os.path.relpath(os.path.dirname(os.path.abspath(__file__))) + "/../../../.."
         )
@@ -155,7 +158,7 @@ def test_linter_success(linter, test_self):
         "REPORT_OUTPUT_FOLDER": tmp_report_folder,
         "LOG_LEVEL": "DEBUG",
         "ENABLE_LINTERS": linter.name,
-        "PRINT_ALL_FILES": True
+        "PRINT_ALL_FILES": True,
     }
     if linter.lint_all_other_linters_files is not False:
         env_vars["ENABLE_LINTERS"] += ",JAVASCRIPT_ES"
@@ -298,8 +301,8 @@ def test_get_linter_version(linter, test_self):
     )
     # Write in linter-versions.json
     root_dir = (
-        "/tmp/lint"
-        if os.path.isdir("/tmp/lint")
+        DEFAULT_DOCKER_WORKSPACE_DIR
+        if os.path.isdir(DEFAULT_DOCKER_WORKSPACE_DIR)
         else os.path.relpath(
             os.path.relpath(os.path.dirname(os.path.abspath(__file__))) + "/../../../.."
         )
@@ -363,8 +366,8 @@ def test_get_linter_help(linter, test_self):
     )
     # Write in linter-helps.json
     root_dir = (
-        "/tmp/lint"
-        if os.path.isdir("/tmp/lint")
+        DEFAULT_DOCKER_WORKSPACE_DIR
+        if os.path.isdir(DEFAULT_DOCKER_WORKSPACE_DIR)
         else os.path.relpath(
             os.path.relpath(os.path.dirname(os.path.abspath(__file__))) + "/../../../.."
         )
@@ -413,10 +416,10 @@ def test_linter_report_tap(linter, test_self):
         f"expected-{linter.descriptor_id}.tap",
     ] + reports_with_extension
     for file_nm in list(dict.fromkeys(possible_reports)):
-        if os.path.isfile(f"{workspace}{os.path.sep}{DEFAULT_REPORT_FOLDER_NAME}{os.path.sep}{file_nm}"):
-            expected_file_name = (
-                f"{workspace}{os.path.sep}{DEFAULT_REPORT_FOLDER_NAME}{os.path.sep}{file_nm}"
-            )
+        if os.path.isfile(
+            f"{workspace}{os.path.sep}{DEFAULT_REPORT_FOLDER_NAME}{os.path.sep}{file_nm}"
+        ):
+            expected_file_name = f"{workspace}{os.path.sep}{DEFAULT_REPORT_FOLDER_NAME}{os.path.sep}{file_nm}"
     if expected_file_name == "":
         raise unittest.SkipTest(
             f"Expected report not defined in {workspace}{os.path.sep}{DEFAULT_REPORT_FOLDER_NAME}"
@@ -482,11 +485,14 @@ def test_linter_report_tap(linter, test_self):
                     f"remain perfectly identical :)"
                 )
 
+
 # Test that the linter provides a SARIF output if it is configured like that
 def test_linter_report_sarif(linter, test_self):
-    if linter.disabled is True or "all" in getattr(
-        linter, "descriptor_flavors_exclude", []
-    ) or linter.can_output_sarif is False:
+    if (
+        linter.disabled is True
+        or "all" in getattr(linter, "descriptor_flavors_exclude", [])
+        or linter.can_output_sarif is False
+    ):
         raise unittest.SkipTest("SARIF is not configured for this linter")
     test_folder = linter.test_folder
     workspace = config.get("DEFAULT_WORKSPACE") + os.path.sep + test_folder
@@ -499,7 +505,7 @@ def test_linter_report_sarif(linter, test_self):
         "REPORT_OUTPUT_FOLDER": tmp_report_folder,
         "ENABLE_LINTERS": linter.name,
         "LOG_LEVEL": "DEBUG",
-        "LOG_FILE": 'megalinter.log'
+        "LOG_FILE": "megalinter.log",
     }
     env_vars.update(linter.test_variables)
     mega_linter, _output = call_mega_linter(env_vars)
@@ -511,7 +517,8 @@ def test_linter_report_sarif(linter, test_self):
         f"{tmp_report_folder}{os.path.sep}sarif{os.path.sep}{linter.name}.sarif"
     )
     test_self.assertTrue(
-        os.path.isfile(tmp_sarif_file_name), f"SARIF report not found {tmp_sarif_file_name}"
+        os.path.isfile(tmp_sarif_file_name),
+        f"SARIF report not found {tmp_sarif_file_name}",
     )